+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/vue
 
+## 0.0.55
+
+### Patch Changes
+
+- caedcda: feat (ai/ui): add setData helper to useChat
+
+## 0.0.54
+
+### Patch Changes
+
+- b5f577e: fix (ui/vue): prevent infinite looping steps
+
 ## 0.0.53
 
 ### Patch Changes